Adding parameter randomization toggle
Added toggle to randomize colors and scale of fireworks. This required changing some of the logic for how the fireworks are created so that values would remain consistent as more fireworks are created while previous ones are still animating. This would have needed to change later anyway since we will need to keep track of unique values per firework for the editor.