Adding firework recording and free mode functionality
Adding functionality to allow fireworks to be recorded so they will play on a loop on the timeline, or be created in free mode (not recorded, timing is not bound to timeline, launch on click instead of explode on click). Also improved firework positioning so that the terminal (explode) point is based on the screen space creation position and not the 3D space position. This allows different window dimensions / device orientations to see all fireworks within their window.