Adding url getter and searchparams for url manipulation
Added initial url and searchparams vars. Global settings for looping and night sky should be checked / set as soon as the page is loaded. Fireworks each should add an additional encoded string to the search params and if fireworks are present in the url when the page is loaded, the site should switch to playback mode.