Adding first url search param
Added a search param for looping. The site checks for the presence of the param on load, and if it returns a value it applies that param to the looping ui and value. If it does not return a value, it applies the default "l=1" looping state param to the url.