Added a warning popup if not all fireworks are valid
If not all fireworks loaded from the url param are considered valid, a popup will appear and the firework(s) in question will be dropped from the url. If The user accepts, they can watch the show without the token that weren't accepted; if they cancel, they go to default edit mode with a blank slate.