Adding functionality for firework aspect ratio
Adding functionality to record window aspect ratio when a firework is created so that the aspect ratio of the fireworks can be preserved when viewing the show on a different screen type. This will make the show appear overall more confined on screens of different size, but will eliminate the stretch and squash issue.